Move skilltree implementation to a new file
