+++ src.cmake
@@ -193,6 +193,7 @@ set(SGAMELIST
     ${GAMELOGIC_DIR}/sgame/sg_bot_parse.cpp
     ${GAMELOGIC_DIR}/sgame/sg_bot_parse.h
     ${GAMELOGIC_DIR}/sgame/sg_bot_public.h
+    ${GAMELOGIC_DIR}/sgame/sg_bot_skilltree.cpp
     ${GAMELOGIC_DIR}/sgame/sg_bot_util.cpp
     ${GAMELOGIC_DIR}/sgame/sg_bot_util.h
     ${GAMELOGIC_DIR}/sgame/sg_buildable.cpp
